Little fixes

+++ src/main/java/com/ikubinfo/primefaces/service/impl/RoomCategoryServiceImpl.java
@@ -1,9 +1,9 @@
 package com.ikubinfo.primefaces.service.impl;
 
 import com.ikubinfo.primefaces.model.RoomCategory;
-import com.ikubinfo.primefaces.repository.BookingRepository;
 import com.ikubinfo.primefaces.repository.RoomCategoryRepository;
 import com.ikubinfo.primefaces.service.RoomCategoryService;
+import com.ikubinfo.primefaces.service.exceptions.CategoryInUseException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Service;
@@ -45,7 +45,18 @@ public class RoomCategoryServiceImpl implements RoomCategoryService {
     }
 
     @Override
-    public void delete(RoomCategory roomCategory) {
-        categoryRepository.deleteCategory(roomCategory);
+    public void delete(RoomCategory roomCategory) throws CategoryInUseException {
+        if (categoryRepository.isCategoryInUse(roomCategory)) {
+            throw new CategoryInUseException(" Cannot delete this category because it is already in use. ");
+        } else {
+            categoryRepository.deleteCategory(roomCategory);
+        }
+
+    }
+
+
+    @Override
+    public boolean isCategoryInUse(RoomCategory roomCategory) {
+        return categoryRepository.isCategoryInUse(roomCategory);
     }
 }
